Fixed some performance issues

+++ proactive-p2p-structured/proactive-p2p-structured-configuration/src/main/java/org/objectweb/proactive/extensions/p2p/structured/configuration/P2PStructuredProperties.java
@@ -41,7 +41,7 @@ public class P2PStructuredProperties {
             new PropertyBoolean("enable.benchmarks.information", false);
 
     public static final PropertyByte CAN_NB_DIMENSIONS = new PropertyByte(
-            "can.nb.dimensions", Byte.valueOf((byte) 3));
+            "can.nb.dimensions", (byte) 3);
 
     /**
      * This property defines the minimum lower bound associated to a CAN
